VTCCA-1027 change Ids to Long

+++ app/models/APIGroupAccountSubmission.scala
@@ -20,7 +20,7 @@ import java.time.Instant
 
 import play.api.libs.json.Json
 
-case class APIGroupAccountSubmission(governmentGatewayGroupId: String, organisationName: String, addressUnitId: Int,
+case class APIGroupAccountSubmission(governmentGatewayGroupId: String, organisationName: String, addressUnitId: Long,
                                      organisationEmailAddress: String, organisationTelephoneNumber: String,
                                      representativeFlag: Boolean,
                                      representativeCode: Option[Int], effectiveFrom: Instant,
